requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
boto3==1.26.155
botocore==1.29.155
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-storages==1.13.2
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
magic-filter==1.0.9
multidict==6.0.4
mysql-connector-python==8.0.33
mysqlclient==2.1.1
protobuf==3.20.3
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
websockets==11.0.3
yarl==1.9.2
